chore(swiper): split demo
