+++ src/packages/__VUE/swiper/demo/index.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script setup lang="ts">
-import { useTranslate } from '../../../utils';
+import { useTranslate } from '@/sites/utils';
 import Basic from './basic.vue';
 import Async from './async.vue';
 import Dynamic from './dynamic.vue';
